feat(payment): refactor
